Move test to launch folder, more helpful error msg, check res=None case
